refactor(web): use one global importer
