+++ apps/web/src/components/auth/Reset.tsx
@@ -2,7 +2,7 @@ import { useState } from 'react';
 import { SubmitHandler, FieldValues } from 'react-hook-form';
 import { toast } from 'sonner';
 
-import Form from '../global/Form';
+import { Form } from '@/components/global';
 
 import { resetInputs } from '@/lib/input';
 import { useAuth } from '@/providers/AuthProvider';
